requirements.txt
torch
sacred
pyyaml
ruamel.yaml
munch
pymongo
numpy
omegaconf